simplify module structure. codict gets preferred if compiled in __init__
